Update configure and build for MRtrix3Tissue user experience
As with set_path, this should avoid confusion for users, so it's clear for them that they are interacting with the MRtrix3Tissue fork. As with set_path, this is made clear at all times, and the relation to MRtrix3 is emphasised at a few key points. For optional assistance with or bug reporting on the configure script, they are also kindly referred to the MRtrix3 community: problems or bugs with this script are highly likely to equally apply to MRtrix3 and MRtrix3Tissue, so this would be beneficial to the community as a whole. This may also allow for bug fixing in MRtrix3, should the opportunity arise.