Update package_mrtrix for MRtrix3Tissue user experience
Similar to set_path, build and configure in earlier commits. Contemplated renaming this script accordingly, but decided to keep the name to avoid confusion. From the output folder name itself, as well as the command line feedback during the execution of the script, it should be obvious what the different context is.