Improve dwi2response dhollander and temporarily add dholl_old
Use lmax=6 eWM FOD peak orientation for signal rotation to estimate WM response function to improve dwi2response dhollander and make it less tensor dependent.

Also reintroduce the old algorithm as dholl_old, but make it clear this is now deprecated and will only be present temporarily, in a limited number of releases. It will be removed again eventually.