Change default non-DEC ODF rendering colour to blue
Most frequent usecase is to overlay a WM-FOD render on a TEC map: the best way to show this in a user friendly and consistent way is by colouring the FODs in a shade of blue, a bit lighter and softer than pure blue though. For other renders (e.g., raw DWI signal profiles and such), this colour also works better than the previous yellow: the shading shows in a visibly more comfortable range, so the cues provided by the shading about the 3D structure of the profile is also more effective.